src/quantum_sim.py
import cirq
import numpy as np

class QuantumCellSimulator:
    def __init__(self):
        self.qubits = [cirq.GridQubit(0, i) for i in range(4)]
        
    def create_reprogramming_circuit(self, target_cell_type):
        """Create quantum circuit for cell reprogramming simulation."""
        if len(target_cell_type) != 4:
            raise ValueError("Target cell type must have 4 parameters")
            
        circuit = cirq.Circuit()
        
        # Initialize quantum state
        circuit.append(cirq.H.on_each(self.qubits))
        
        # Entangle qubits to represent cell state interactions
        for i in range(len(self.qubits) - 1):
            circuit.append(cirq.CNOT(self.qubits[i], self.qubits[i + 1]))
            
        # Apply rotation gates based on target cell type
        for i, qubit in enumerate(self.qubits):
            circuit.append(cirq.rx(rads=np.pi * target_cell_type[i])(qubit))
            
        # Measurement
        circuit.append(cirq.measure(*self.qubits, key='result'))
        
        return circuit
    
    def simulate_reprogramming(self, target_cell_type):
        """Simulate cell reprogramming process with intermediate states."""
        if len(target_cell_type) != 4:
            raise ValueError("Target cell type must have 4 parameters")
            
        circuit = self.create_reprogramming_circuit(target_cell_type)
        simulator = cirq.Simulator()
        
        # Get intermediate states
        intermediate_states = []
        for i, moment in enumerate(circuit):
            if i < len(circuit) - 1:  # Skip final measurement for state calculation
                temp_circuit = circuit[:i+1]
                result = simulator.simulate(temp_circuit)
                intermediate_states.append({
                    'step': i,
                    'operation': str(moment),
                    'state_vector': result.final_state_vector
                })
        
        # Get final measurement results
        result = simulator.run(circuit, repetitions=100)
        final_results = result.histogram(key='result')
        
        return {
            'final_results': final_results,
            'intermediate_states': intermediate_states,
            'circuit_diagram': str(circuit),
            'qubit_count': len(self.qubits),
            'success_metric': sum(v for k, v in final_results.items() if k > 0.5) / 100
        }
    
    def optimize_transformation(self, initial_state, target_state, steps=10):
        """Optimize cell transformation pathway with diagnostics."""
        circuit = cirq.Circuit()
        
        # Create superposition of possible transformation pathways
        circuit.append(cirq.H.on_each(self.qubits))
        
        # Add parameterized gates for optimization
        optimization_steps = []
        for step in range(steps):
            step_operations = []
            for i in range(len(self.qubits)):
                gate = cirq.rx(rads=np.pi/2)(self.qubits[i])
                circuit.append(gate)
                step_operations.append(str(gate))
                if i < len(self.qubits) - 1:
                    gate = cirq.CNOT(self.qubits[i], self.qubits[i + 1])
                    circuit.append(gate)
                    step_operations.append(str(gate))
            
            optimization_steps.append({
                'step': step,
                'operations': step_operations
            })
        
        return {
            'circuit': circuit,
            'optimization_steps': optimization_steps,
            'total_gates': len(list(circuit.all_operations())),
            'circuit_depth': len(circuit),
            'initial_state': initial_state.tolist(),
            'target_state': target_state.tolist()
        }
